+++ src/ts/Adders/Home/HomeLinksAdder.ts
@@ -8,9 +8,11 @@ import homeLinksSectionTemplate from '@/templates/homeLinksSectionTemplate'
 import linksIcons from '@/static/linksIcons'
 
 const DEFAULT_ICON_NAME = 'link-white.png'
+const EVENT_LINKS = ['/conferences', '/cal.php']
 
 export default class HomeLinksAdder implements AdderInterface {
     private sidebarElem: Element
+    private recommendedLinks: HTMLDivElement[] = []
 
     public constructor() {
         this.sidebarElem = document.querySelector(conf.selectors.home.rightSidebar)!
@@ -21,8 +23,11 @@ export default class HomeLinksAdder implements AdderInterface {
             return
         }
 
-        this.addLinksSection(this.getRecommendedLinks(), 'Recommended Links')
-        this.addLinksSection(this.getSocialLinks(), 'Social Links')
+        this.recommendedLinks = this.getRecommendedLinksFromPage()
+
+        this.addLinksSection(this.getSocialLinks(), 'Social')
+        this.addLinksSection(this.getRecommendedLinks(), 'Recommended')
+        this.addLinksSection(this.getEventsLinks(), 'Events')
 
         this.removeInitialLinks()
     }
@@ -31,7 +36,9 @@ export default class HomeLinksAdder implements AdderInterface {
         const innerElem = this.sidebarElem.querySelector('.inner')
 
         if (!innerElem) {
-            console.warn('[PHP Revival]: Could not find the inner element in the sidebar to remove it')
+            console.warn(
+                '[PHP Revival]: Could not find the inner element in the sidebar to remove it',
+            )
             return
         }
 
@@ -51,11 +58,13 @@ export default class HomeLinksAdder implements AdderInterface {
     }
 
     private displayLinks(sidebarSection: HTMLElement): void {
-        setTimeout(() => sidebarSection.style.opacity = '1', 300)
+        setTimeout(() => (sidebarSection.style.opacity = '1'), 300)
     }
 
     private getRecommendedLinks(): HTMLElement[] {
-        const linkElements: HTMLElement[] = this.getRecommendedLinksFromPage()
+        const linkElements = this.recommendedLinks.filter(
+            this.filterRecommendedLinks,
+        )
 
         const additionalLinks = recommendedLinks.map(link => homeLinkTemplate(link))
 
@@ -64,9 +73,34 @@ export default class HomeLinksAdder implements AdderInterface {
         return linkElements
     }
 
-    private getSocialLinks(): HTMLElement[] {
-        const linkElements: HTMLElement[] = this.getSocialLinksFromPage()
+    private getEventsLinks(): HTMLElement[] {
+        return this.recommendedLinks.filter(this.filterEventsLinks)
+    }
 
+    private filterRecommendedLinks(link: HTMLDivElement): boolean {
+        return !EVENT_LINKS.some(href => {
+            if (!link.dataset.url) {
+                console.warn('[PHP Revival]: No dataset.url found in the link')
+                return false
+            }
+
+            return link.dataset.url.includes(href)
+        })
+    }
+
+    private filterEventsLinks(link: HTMLDivElement): boolean {
+        return EVENT_LINKS.some(href => {
+            if (!link.dataset.url) {
+                console.warn('[PHP Revival]: No dataset.url found in the link')
+                return false
+            }
+
+            return link.dataset.url.includes(href)
+        })
+    }
+
+    private getSocialLinks(): HTMLDivElement[] {
+        const linkElements = this.getSocialLinksFromPage()
         const additionalLinks = socialLinks.map(link => homeLinkTemplate(link))
 
         linkElements.push(...additionalLinks)
@@ -74,41 +108,51 @@ export default class HomeLinksAdder implements AdderInterface {
         return linkElements
     }
 
-    private getRecommendedLinksFromPage(): HTMLElement[] {
+    private getRecommendedLinksFromPage(): HTMLDivElement[] {
         const links: HomeLink[] = []
 
         const elements = this.sidebarElem.querySelectorAll('.inner > .panel')
         const panels = Array.from(elements)
 
         for (const panel of panels) {
-            if (this.childrenHaveLinks(panel.children)) {
-                const link = this.getLinkFromChildren(panel.children)
-
-                if (link) {
-                    links.push(link)
-                }
-
+            if (!this.childrenHaveLinks(panel.children)) {
                 continue
             }
+
+            const link = this.getLinkFromChildren(panel.children)
+
+            if (!link) {
+                continue
+            }
+
+            const nestedList = panel.querySelector('.body > ul')
+
+            if (nestedList) {
+                link.nestedList = nestedList as HTMLUListElement
+            }
+
+            links.push(link)
         }
 
         return links.map(link => homeLinkTemplate(link))
     }
 
-    private getSocialLinksFromPage(): HTMLElement[] {
+    private getSocialLinksFromPage(): HTMLDivElement[] {
         const links: HomeLink[] = []
 
-        const elements = this.sidebarElem.querySelectorAll('.inner > .social-media ul li a')
+        const selector = '.inner > .social-media ul li a'
+        const elements = this.sidebarElem.querySelectorAll(selector)
+
         const anchorTags = Array.from(elements)
 
         for (const anchor of anchorTags) {
             if (anchor.tagName !== 'A') {
-                console.warn(`[PHP Revival]: Expected an anchor tag, got ${anchor.tagName}`)
+                const msg = `[PHP Revival]: Expected an anchor tag, got ${anchor.tagName}`
+                console.warn(msg)
                 continue
             }
 
             const element = anchor as HTMLAnchorElement
-
             const title = element.innerText!.trim()
 
             links.push({