+++ src/ts/Adders/SearchIconAdder.ts
@@ -7,7 +7,9 @@ export default class SearchIconAdder implements AdderInterface {
         const target = document.querySelector(conf.selectors.targetForSearchIcon)
 
         if (!target) {
-            console.warn(`[PHP Revival]: Could not find target for search icon: ${conf.selectors.targetForSearchIcon}`)
+            console.warn(
+                `[PHP Revival]: Could not find target for search icon: ${conf.selectors.targetForSearchIcon}`,
+            )
             return
         }