+++ src/ts/Modifiers/CodeSampleModifier.ts
@@ -6,8 +6,12 @@ export default class CodeSampleModifier implements Modifier {
     private staticClasses: NodeListOf<HTMLDivElement>
 
     public constructor() {
-        this.codeAreas = document.querySelectorAll<HTMLDivElement>(conf.selectors.codeExamples)
-        this.staticClasses = document.querySelectorAll<HTMLDivElement>(conf.selectors.docs.classMethods)
+        this.codeAreas = document.querySelectorAll<HTMLDivElement>(
+            conf.selectors.codeExamples,
+        )
+        this.staticClasses = document.querySelectorAll<HTMLDivElement>(
+            conf.selectors.docs.classMethods,
+        )
     }
 
     public modify(): void {
@@ -27,7 +31,10 @@ export default class CodeSampleModifier implements Modifier {
                 .replace(/\}<br>(\w|\$)/, '}<br><br>$1')
                 .replace(/([;]+)<br>([a-z\/])/, ';<br><br>$2')
                 .replace(/function\(/, 'function (')
-                .replace(/([;}])<br>([&nbsp;\s]+ (if|while|foreach|for|which))/, '$1<br><br>$2')
+                .replace(
+                    /([;}])<br>([&nbsp;\s]+ (if|while|foreach|for|which))/,
+                    '$1<br><br>$2',
+                )
                 .replace(/(if|while|foreach|for|which)\(/, '$1 (')
 
             area.innerHTML = newHtml