+++ src/ts/Modifiers/LogoModifier.ts
@@ -4,7 +4,9 @@ import conf from '@/conf'
 
 export default class LogoModifier implements Modifier {
     public modify(): void {
-        const logo = document.querySelector<HTMLImageElement>(conf.selectors.home.logo)
+        const logo = document.querySelector<HTMLImageElement>(
+            conf.selectors.home.logo,
+        )
 
         if (!logo) {
             return