+++ src/ts/types/index.ts
@@ -2,6 +2,7 @@ export type HomeLink = {
     href: string
     iconName: string
     title: string
+    nestedList?: HTMLUListElement // ul > li > a
 }
 
 export type RecommendedVideo = {